+++ lib/attachToGulp.js
@@ -25,6 +25,11 @@ const attachToGulp = (gulp, opts) =>{
 			gulp: gulp
 		});
 
+		gulp.on('stop', function(){
+
+            Deploy.startWatchers();
+        });
+
 		return Deploy.prepare()
 			.then(() =>{
 				return Deploy.composer();
@@ -44,9 +49,6 @@ const attachToGulp = (gulp, opts) =>{
 			.then(() =>{
 				return Deploy.js();
 			})
-			.then(() =>{
-				return Deploy.startWatchers();
-			})
 	});
 
 	gulp.task('deploy.servicelayer', () =>{
@@ -55,6 +57,12 @@ const attachToGulp = (gulp, opts) =>{
 			to: opts.buildDir,
 			gulp: gulp
 		});
+
+        gulp.on('stop', function(){
+
+        	process.env.PRTL_DISABLED_WATCHERS = process.env.PRTL_DISABLED_WATCHERS || 'js,scss';
+            Deploy.startWatchers();
+        });
 
 		return Deploy.prepare()
 			.then(() =>{
