+++ lib/deploy.js
@@ -2,6 +2,7 @@
 
 const env = require('./private/env');
 const log = require('@studyportals/node-log');
+const sass = require('./private/sass');
 
 /**
  * @type {Deploy}
@@ -130,7 +131,7 @@ class Deploy {
 	 */
 	sass(){
 
-		return require('./private/sass').compile({
+		return sass.compile({
 			'base': this.opts.to,
 			'from': [
 				`${this.opts.to}/**/*.scss`,
@@ -179,7 +180,10 @@ class Deploy {
 	 * - js
 	 * - scss
 	 * - file
-	 *
+     *
+     * Watchers can be disabled by providing a comma separated list in the env var `PRTL_DISABLED_WATCHERS`
+     * For instance `PRTL_DISABLED_WATCHERS=file` will enable js and scss watchers but disables the file watcher.
+     *
 	 * @return {undefined}
 	 */
 	startWatchers(){
@@ -198,8 +202,6 @@ class Deploy {
             // All disabled, nothing to do.
             return;
         }
-
-        log.info(`Starting watchers, excluding: ${disabledWatchers.join(', ')}`);
 
         const chokidar = require('chokidar');
         const gulp = this.opts.gulp;
@@ -221,17 +223,25 @@ class Deploy {
 
         if(disabledWatchers.indexOf('js') !== -1){
 
+            log.info(`Watching js files is disabled.`);
             ignored.push('**/*.js');
         }
 
         if(disabledWatchers.indexOf('scss') !== -1){
 
+            log.info(`Watching scss files is disabled.`);
+            ignored.push('**/*.scss');
+        }
+
+        if(disabledWatchers.indexOf('file') !== -1){
+
+            log.info(`Watching files is disabled.`);
             ignored.push('**/*.scss');
         }
 
         const compileSass = function(filePath){
 
-            return require('./private/sass').compile({
+            return sass.compile({
                 base: self.opts.from,
                 from: filePath,
                 to: self.opts.to,
@@ -254,25 +264,14 @@ class Deploy {
                     break;
                 case '.scss':
 
-                    // Basename without extenstion
-                    let basename = path.basename(filePath).split('.').shift();
+                    if(path.basename(filePath).startsWith('_')){
 
-                    if(basename.startsWith('_')){
-
-                        // Strip off the _
-                        basename = basename.substring(1);
-
-                        require('./private/sass-references')()
+                        sass.getReferences()
                             .then(ref => {
 
                                 const dependencies = ref[path.resolve(filePath)];
                                 compileSass(Object.keys(dependencies));
                             });
-
-                        // console.log(`FIND FILES: ${references[path.resolve(filePath)]}`);
-
-                        // TODO: handle dependencies.
-                        // XXX: if starts with _, find usages and compile those
                     }
                     else{
 
@@ -313,12 +312,13 @@ class Deploy {
 
         chokidar.watch('.', {
             ignoreInitial: true,
-            awaitWriteFinish: true,
             ignored
         })
 		.on('add', change)
 		.on('change', change)
 		.on('unlink', unlink);
+
+        log.info(`Watchers started, use 'CTRL+c' to stop the watchers.`);
 	}
 }
