+++ lib/private/config.js
@@ -7,20 +7,20 @@ try{
 
 		process.env[key] = Config[key];
 	});
-	console.log(`Using 'Config.json'`);
+    require('@studyportals/node-log').debug(`Env vars are overwritten by 'Config.json'`);
 }
 catch(e){
 
-	console.log(`No 'Config.json' found, using default.`);
+	// No 'Config.json' found, using default.
 }
 
 const log = require('@studyportals/node-log');
 
-console.log('PRTL env vars:');
+log.debug('PRTL env vars:');
 Object.keys(process.env).forEach(function(key){
 
 	if(key.startsWith("PRTL")){
 
-		console.log(`${key} => ${process.env[key]}`);
+		log.debug(`${key} => ${process.env[key]}`);
 	}
 });
