+++ lib/private/deploylog.js
@@ -4,7 +4,15 @@ const request = require('request');
 const log = require('@studyportals/node-log');
 const startTime = new Date();
 
+let send = false;
+
 const logState = (state, time, success, extra) => {
+
+    if(send){
+
+        // The deploy either succeeds or fails, the second time it are probably watchers.
+        return;
+    }
 
     log.debug(`logState: ${state}. ${time}, ${success}, ${extra}`);
 
@@ -21,6 +29,8 @@ const logState = (state, time, success, extra) => {
     const pjson = require('../../package.json');
     const pjsonMain = require(path.resolve(process.cwd(), 'package.json'));
     const cwd = process.cwd();
+
+    send = true;
 
     try {
         request.post(process.env.PRTL_DEPLOYLOG_ENDPOINT, {
